pyproject.toml
[project]
name = "news-bot"
version = "0.1.0"
description = "Interactive terminal news reader for East Asian news sources"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0.10",
    "trafilatura>=1.6.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.1",
]

[project.scripts]
news = "news_bot.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/news_bot"]

[dependency-groups]
dev = []
